move reactflow code into lib
