--- ui/src/features/Editor/components/Canvas/components/index.ts
@@ -1,6 +1,0 @@
-export * from "./Nodes";
-export {
-  default as CustomConnectionLine,
-  connectionLineStyle,
-} from "./CustomConnectionLine";
-export { default as CustomEdge, edgeTypes } from "./CustomEdge";
