+++ ui/src/lib/reactFlow/edgeTypes/index.ts
@@ -1,0 +1,7 @@
+import { EdgeTypes } from "@xyflow/react";
+
+import CustomEdge from "./CustomEdge";
+
+export const edgeTypes: EdgeTypes = {
+  default: CustomEdge,
+};
