+++ ui/src/lib/reactFlow/nodeTypes/NoteNode/index.tsx
