+++ ui/src/lib/reactFlow/nodeTypes/WriterNode.tsx
