+++ ui/src/lib/reactFlow/nodeTypes/index.ts
@@ -1,8 +1,6 @@
-import { NodeTypes } from "@xyflow/react";
+import type { NodeTypes } from "@xyflow/react";
 
 import BatchNode from "./BatchNode";
-import EntranceNode from "./EntranceNode";
-import ExitNode from "./ExitNode";
 import NoteNode from "./NoteNode";
 import ReaderNode from "./ReaderNode";
 import SubworkflowNode from "./SubworkflowNode";
@@ -16,6 +14,4 @@ export const nodeTypes: NodeTypes = {
   batch: BatchNode,
   note: NoteNode,
   subworkflow: SubworkflowNode,
-  entrance: EntranceNode,
-  exit: ExitNode,
 };
