+++ ui/src/lib/reactFlow/useEdges.ts
@@ -9,11 +9,12 @@ import { useCallback } from "react";
 import { Edge } from "@flow/types";
 
 type Props = {
+  edges: Edge[];
   onEdgesAdd: (newEdges: Edge[]) => void;
   onEdgesChange: (changes: EdgeChange[]) => void;
 };
 
-export default ({ onEdgesAdd, onEdgesChange }: Props) => {
+export default ({ edges, onEdgesAdd, onEdgesChange }: Props) => {
   const handleEdgesChange: OnEdgesChange<Edge> = useCallback(
     (changes) => onEdgesChange(changes),
     [onEdgesChange],
@@ -23,6 +24,7 @@ export default ({ onEdgesAdd, onEdgesChange }: Props) => {
     (connection) => {
       // Reference: https://github.com/xyflow/xyflow/blob/043c8120ace53b562c0b3ec8194ccdef64ccad0e/packages/system/src/utils/edges/general.ts#L79
       const edgeId = `xy-edge__${connection.source}${connection.sourceHandle || ""}-${connection.target}${connection.targetHandle || ""}`;
+      if (edges.find((e) => e.id === edgeId)) return;
       onEdgesAdd([
         {
           id: edgeId,
@@ -30,7 +32,7 @@ export default ({ onEdgesAdd, onEdgesChange }: Props) => {
         },
       ]);
     },
-    [onEdgesAdd],
+    [edges, onEdgesAdd],
   );
 
   const handleReconnect: OnReconnect = useCallback(
