+++ ui/src/lib/reactFlow/utils.ts
@@ -1,0 +1,7 @@
+import { Connection } from "@xyflow/react";
+
+import { Edge } from "@flow/types";
+
+export function isValidConnection(connection: Connection | Edge): boolean {
+  return connection.source !== connection.target;
+}
